Remove git diff from CommandRunner's reponsibility
