+++ src/main/java/reposense/git/GitDiff.java
@@ -1,0 +1,17 @@
+package reposense.git;
+
+import static reposense.system.CommandRunner.runCommand;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+public class GitDiff {
+
+    /**
+     * Returns the git diff result of the current commit compared to {@code lastCommitHash}, without any context.
+     */
+    public static String diffCommit(String root, String lastCommitHash) {
+        Path rootPath = Paths.get(root);
+        return runCommand(rootPath, "git diff -U0 " + lastCommitHash);
+    }
+}
